[#166734445] remove services from static container, rename container file to config
