+++ src/server.ts
@@ -7,28 +7,28 @@ import * as https from "https";
 import { NodeEnvironmentEnum } from "italia-ts-commons/lib/environment";
 import { newApp } from "./app";
 import {
+  ALLOW_NOTIFY_IP_SOURCE_RANGE,
+  ALLOW_PAGOPA_IP_SOURCE_RANGE,
   API_BASE_PATH,
   AUTHENTICATION_BASE_PATH,
-  container,
+  ENV,
   PAGOPA_BASE_PATH,
   SAML_CERT,
-  SAML_KEY
-} from "./container";
+  SAML_KEY,
+  SERVER_PORT
+} from "./config";
 import { log } from "./utils/logger";
 
-const port = container.resolve("serverPort");
-const env = container.resolve("env");
-
-const authenticationBasePath = container.resolve(AUTHENTICATION_BASE_PATH);
-const APIBasePath = container.resolve(API_BASE_PATH);
-const PagoPABasePath = container.resolve(PAGOPA_BASE_PATH);
+const authenticationBasePath = AUTHENTICATION_BASE_PATH;
+const APIBasePath = API_BASE_PATH;
+const PagoPABasePath = PAGOPA_BASE_PATH;
 
 // tslint:disable-next-line: no-let
 let server: http.Server | https.Server;
 newApp(
-  env,
-  container.resolve("allowNotifyIPSourceRange"),
-  container.resolve("allowPagoPAIPSourceRange"),
+  ENV,
+  ALLOW_NOTIFY_IP_SOURCE_RANGE,
+  ALLOW_PAGOPA_IP_SOURCE_RANGE,
   authenticationBasePath,
   APIBasePath,
   PagoPABasePath
@@ -36,15 +36,14 @@ newApp(
   .then(app => {
     // In test and production environments the HTTPS is terminated by the Kubernetes Ingress controller. In dev we don't use
     // Kubernetes so the proxy has to run on HTTPS to behave correctly.
-    if (env === NodeEnvironmentEnum.DEVELOPMENT) {
-      const samlKey = container.resolve(SAML_KEY);
-      const options = { key: samlKey, cert: SAML_CERT };
+    if (ENV === NodeEnvironmentEnum.DEVELOPMENT) {
+      const options = { key: SAML_KEY, cert: SAML_CERT };
       server = https.createServer(options, app).listen(443, () => {
         log.info("Listening on port 443");
       });
     } else {
-      server = http.createServer(app).listen(port, () => {
-        log.info("Listening on port %d", port);
+      server = http.createServer(app).listen(SERVER_PORT, () => {
+        log.info("Listening on port %d", SERVER_PORT);
       });
     }
     server.on("close", () => {
